Initialize and Set-up Project
Name: Clock of clocks
Description: An digital made of multiple clocks

This commit includes the core functionality of the project.
The plan is to have a 8x15 grid (which is already implemented in the
`containers/clock_mesh.dart` and is filled with randomly colored
rectangles) - The color changes every second, because the widget is rebuilt
every time the timer updates in `clock.dart`.