Test [ClockState] action and listeners
- All the analog clocks are now listening for changes in the
global [ClockState] and update whenever it's necessary (when
their object has been modified). Problem is that even if we
update one single ClockModel every singled [AnalogClock] will
be notified and re-built. As a solution I'm looking towards
implementing PropertyChangeNotifier, if that won't work
there's a heavier alternative: compare the new object
with the old one every single time the [AnalogClock] widget
is notified, if they are the same, don't trigger re-build
(updateShouldNotify() or didChangeDependencies()).