Use property_change_notifier library
- Experiment with the property_change_notifier library.
This library is an alternative to the ChangeNotifier
and has some great features that allow listeners of
PropertyChangeProvider to observe specific properties.
- This indeed fixes the problem of notifying all widgets.
This implementation only notifies single clock widgets
(on each tap it generates a random number (0..119),
updates its state and notifies the corresponding widget).

NOTE: Problem is that there's no way of triggering changes
to multiple listeners of the same typy at the same time.
And that is kind of fundamental for this project.