Add Widespread Clock Arrangements
In simmilar vein with (e81ce99) the [clockArrangements]
use the [arrangeClockGroups()] helper to arrange any widespread
clock arrangements. This commit only includes a default arrangement,
but in the future there's a high chance more will be added.

As the doc. comments state, these don't cover digit representation,
but they could have a certain arrangement which reflects a certain time.