Fix arrangers. Make copies instead of references.
This issue came to light upon trying to change the color
of a group of [ClockHandModel]'s of a certain digit. Since
the arrangers were simply creating references to the default
arrangements, whenever a change was made for a specific digit
it was also reflected in other digits (based on implementation).